LPS-104351 Update tests
