+++ modules/dxp/apps/portal-workflow/portal-workflow-metrics-web/test/shared/components/results-bar/ResultsBar.es.js
@@ -12,19 +12,23 @@
 import React from 'react';
 import renderer from 'react-test-renderer';
 
-import ProcessListItem from '../../../src/main/resources/META-INF/resources/js/components/process-list/ProcessListItem.es';
-import {MockRouter as Router} from '../../mock/MockRouter.es';
+import ResultsBar from '../../../../src/main/resources/META-INF/resources/js/shared/components/results-bar/ResultsBar.es';
+import {MockRouter as Router} from '../../../mock/MockRouter.es';
 
-test('Should render component with one list item', () => {
+test('Should render component with one item', () => {
 	const component = renderer.create(
 		<Router>
-			<ProcessListItem
-				id={36401}
-				instanceCount={10}
-				onTimeInstanceCount={5}
-				overdueInstanceCount={5}
-				title="Process test"
-			/>
+			<ResultsBar>
+				<>
+					<ResultsBar.TotalCount search={'test'} totalCount={1} />
+
+					<ResultsBar.Clear
+						page={1}
+						pageSize={5}
+						sort={encodeURIComponent('title:asc')}
+					/>
+				</>
+			</ResultsBar>
 		</Router>
 	);
 
@@ -33,16 +37,20 @@ test('Should render component with one list item', () => {
 	expect(tree).toMatchSnapshot();
 });
 
-test('Should render component with 1 instance count', () => {
+test('Should render component with one list item', () => {
 	const component = renderer.create(
 		<Router>
-			<ProcessListItem
-				id={36401}
-				instanceCount={1}
-				onTimeInstanceCount={5}
-				overdueInstanceCount={5}
-				title="Process test"
-			/>
+			<ResultsBar>
+				<>
+					<ResultsBar.TotalCount search={'test'} totalCount={5} />
+
+					<ResultsBar.Clear
+						page={1}
+						pageSize={5}
+						sort={encodeURIComponent('title:asc')}
+					/>
+				</>
+			</ResultsBar>
 		</Router>
 	);
